Feat: complete user story 3
