+++ app/controllers/bulk_discounts_controller.rb
@@ -23,6 +23,13 @@ class BulkDiscountsController < ApplicationController
     redirect_to merchant_bulk_discounts_path(@merchant) 
   end
 
+  def destroy
+    @merchant = Merchant.find(params[:merchant_id])
+    @bulk_discount = @merchant.bulk_discounts.find(params[:id])
+    @bulk_discount.destroy!
+    redirect_to merchant_bulk_discounts_path
+  end
+
   private
   
   def item_params
