+++ app/views/bulk_discounts/index.html.erb
@@ -1,4 +1,3 @@
-
 <h1>Bulk Discounts Index</h1>
 
 <%= link_to "Create New Discount", new_merchant_bulk_discount_path(@merchant), class: "create-new-discount-link" %>
@@ -11,6 +10,7 @@
         <p>Percentage: <%= bulk_discount.decimal_to_percentage %></p>
         <p>Quantity: <%= bulk_discount.quantity %></p>
         <p><%= link_to "More Info", merchant_bulk_discounts_path(merchant_id: bulk_discount.merchant_id, id: bulk_discount.id) %></p>
+        <p><%= link_to "Delete", merchant_bulk_discount_path(merchant_id: bulk_discount.merchant_id, id: bulk_discount.id), method: :delete, data: { turbo_method: :delete, turbo_confirm: "Are you sure?" } %></p>
       </div>
     <% end %>
   </div>
